+++ src/pages/AboutPage.jsx
@@ -177,7 +177,7 @@ function AboutPage() {
             <FirstAlphabet color='blue'>D</FirstAlphabet>reamers
           </span>
         </GoblinP>
-        <P line='1.5' margin='2rem 0 0 0' size='1.3rem'>
+        <P line='1.5' margin='2rem 0 0 0' size='1rem'>
           FIELD란, ‘Future Industrial Engineering Leaders and Dreamers’ 의 약자로, 미래의 핵심
           리더들이 될 산업 공학도들이 모여 서로의 꿈과 비전, 생각 등을 공유할 수 있는 교류의 장을
           만든다는 목표 아래 모인 &apos;전국 대학생 산업공학도 동아리&apos; 입니다.
